launcher/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
pydantic>=2.11.0,<3.0.0
pyyaml>=6.0.1
aiohttp>=3.11.0
psutil>=6.1.0
jinja2>=3.1.5
anyio>=4.7.0
